+++ packages/deploy-configurations/configs/index.ts
@@ -3,4 +3,5 @@ export { config as goerliConfig } from './goerli.conf'
 export { config as localExtendConfig } from './local-extend.conf'
 export { config as mainnetConfig } from './mainnet.conf'
 export { config as optimismConfig } from './optimism.conf'
+export { config as testConfig } from './test.conf'
 export { getConfigByNetwork } from './utils'
