+++ packages/deploy-configurations/constants/contract-names.local.ts
@@ -80,13 +80,6 @@ export const SERVICE_REGISTRY_NAMES = {
     ERC20_POOL_FACTORY: 'ERC20PoolFactory',
     AJNA_POOL_UTILS_INFO: 'AjnaPoolUtilsInfo',
   },
-  test: {
-    DUMMY_ACTION: 'DummyAction',
-    DUMMY_OPTIONAL_ACTION: 'DummyOptionalAction',
-    DUMMY_SWAP: 'DummySwap',
-    DUMMY_EXCHANGE: 'DummyExchange',
-    SWAP: 'uSwap',
-  },
 } as const
 
 export type AllValues<T> = { [K in keyof T]: T[K] extends object ? AllValues<T[K]> : T[K] }[keyof T]
