+++ packages/deploy-configurations/types/deployment-config/test.ts
@@ -2,7 +2,7 @@ import { Address } from '@deploy-configurations/types/address'
 
 import { ConfigEntry } from './config-entries'
 
-export type TestContractNames = 'DummyAction' | 'DummyOptionalAction' | 'DummyExchange'
+export type TestContractNames = 'DummyAction' | 'DummyOptionalAction' | 'MockExchange'
 
 export type TestConfigEntry = ConfigEntry & {
   name: TestContractNames
