+++ packages/dma-common/constants/index.ts
@@ -10,6 +10,7 @@ export {
   ONE_TEN_THOUSANDTH,
   ONE_THOUSAND,
   TEN,
+  TEN_MILLION,
   TEN_THOUSAND,
   TYPICAL_PRECISION,
   UNUSED_FLASHLOAN_AMOUNT,
