+++ packages/dma-common/constants/numbers.ts
@@ -9,6 +9,7 @@ export const ONE_THOUSAND = new BigNumber(10000)
 export const ONE_TEN_THOUSANDTH = new BigNumber(0.0001)
 export const TEN_THOUSAND = new BigNumber(10000)
 export const MILLION = new BigNumber('1000000')
+export const TEN_MILLION = new BigNumber('10000000')
 export const BILLION = new BigNumber('1000000000')
 export const MAX_UINT =
   '115792089237316195423570985008687907853269984665640564039457584007913129639935'
