+++ packages/dma-common/test-utils/index.ts
@@ -4,7 +4,7 @@ export * from './advance-blocks'
 export * from './advance-time'
 export * from './as-percentage'
 export * from './create-dpm-account'
-export * from './deploy-system'
+export * from './deploy-test-system'
 export * from './dummy-exchange'
 export * from './ensure-wei-format'
 export * from './expect'
