+++ packages/dma-common/types/common.ts
@@ -1,3 +1,4 @@
+import { FakeDAI, FakeWETH, MockExchange } from '@dma-contracts/typechain'
 import { ethers, providers } from 'ethers'
 
 export type UnboxPromise<T> = T extends Promise<infer U> ? U : T
@@ -41,6 +42,24 @@ export interface OneInchSwapResponse extends OneInchBaseResponse {
     value: string
     gasPrice: string
   }
+}
+
+export type FakeRequestEnv = {
+  mockExchange: MockExchange
+  fakeWETH: FakeWETH
+  fakeDAI: FakeDAI
+}
+
+export type OneInchSwapRequest = {
+  fromTokenAddress: string
+  toTokenAddress: string
+  amount: string
+  recipient: string
+  slippage: string
+  protocols?: string[]
+  chainId?: number
+  version?: string
+  fakeRequestEnv?: FakeRequestEnv
 }
 
 export type SwapData = {
