+++ packages/dma-common/utils/swap/calculate-fee.ts
@@ -8,3 +8,10 @@ export function calculateFee(amountWei: BigNumber, fee: number = DEFAULT_FEE): B
     .abs()
     .integerValue(BigNumber.ROUND_DOWN)
 }
+
+export function calculateFeeOnInputAmount(
+  amountWei: BigNumber,
+  fee: number = DEFAULT_FEE,
+): BigNumber {
+  return amountWei.times(fee).div(new BigNumber(FEE_BASE)).abs().integerValue(BigNumber.ROUND_UP)
+}
