+++ packages/dma-contracts/test/unit/actions/common/reentrancy-guard.test.ts
@@ -1,19 +1,18 @@
 import CDPManagerABI from '@abis/external/protocols/maker/dss-cdp-manager.json'
 import { ADDRESSES } from '@deploy-configurations/addresses'
+import { loadContractNames } from '@deploy-configurations/constants'
+import { DeployedSystem } from '@deploy-configurations/types/deployed-system'
 import { Network } from '@deploy-configurations/types/network'
-import { ServiceRegistry } from '@deploy-configurations/utils/wrappers'
-import { CONTRACT_NAMES, OPERATION_NAMES } from '@dma-common/constants'
-import { DeployedSystemInfo, expect, restoreSnapshot } from '@dma-common/test-utils'
-import { RuntimeConfig } from '@dma-common/types/common'
+import { OperationsRegistry as OperationRegistryWrapper } from '@deploy-configurations/utils/wrappers'
+import { expect, restoreSnapshot } from '@dma-common/test-utils'
 import { executeThroughProxy } from '@dma-common/utils/execute'
 import { getLastVault } from '@dma-common/utils/maker'
 import { testBlockNumber } from '@dma-contracts/test/config'
-import { initialiseConfig } from '@dma-contracts/test/fixtures'
 import { ActionFactory, calldataTypes } from '@dma-library'
 import { JsonRpcProvider } from '@ethersproject/providers'
-import { loadFixture } from 'ethereum-waffle'
-import { Signer } from 'ethers'
-import { ethers } from 'hardhat'
+import { DsProxy } from '@typechain'
+import { Signer, utils } from 'ethers'
+import hre, { ethers } from 'hardhat'
 
 const createAction = ActionFactory.create
 
@@ -21,25 +20,42 @@ const createAction = ActionFactory.create
 describe.skip(`Reentrancy guard test | Unit`, async () => {
   let provider: JsonRpcProvider
   let signer: Signer
-  let system: DeployedSystemInfo
-  let registry: ServiceRegistry
-  let config: RuntimeConfig
+  let system: DeployedSystem
+  let userProxy: DsProxy
+  let OPERATION_NAME: string
+  let OpenVaultActionHash: string
+  let operationsRegistry: OperationRegistryWrapper
 
   before(async () => {
-    ;({ config, provider, signer } = await loadFixture(initialiseConfig))
+    const { snapshot } = await restoreSnapshot({ hre, blockNumber: testBlockNumber })
+    system = snapshot.testSystem.deployment.system
+    userProxy = snapshot.testSystem.userProxy
+    provider = snapshot.config.provider
+    signer = snapshot.config.signer
 
-    const { snapshot } = await restoreSnapshot({ config, provider, blockNumber: testBlockNumber })
-    system = snapshot.deployed.system
-    registry = snapshot.deployed.registry
+    operationsRegistry = new OperationRegistryWrapper(
+      snapshot.testSystem.deployment.system.OperationsRegistry.contract.address,
+      signer,
+    )
+
+    // Add new operation with optional Actions
+    const SERVICE_REGISTRY_NAMES = loadContractNames(Network.MAINNET)
+
+    OPERATION_NAME = 'TEST_OPERATION_1'
+    OpenVaultActionHash = utils.keccak256(
+      utils.toUtf8Bytes(SERVICE_REGISTRY_NAMES.maker.OPEN_VAULT),
+    )
+
+    await operationsRegistry.addOp(OPERATION_NAME, [{ hash: OpenVaultActionHash, optional: true }])
   })
 
   afterEach(async () => {
-    await restoreSnapshot({ config, provider, blockNumber: testBlockNumber })
+    await restoreSnapshot({ hre, blockNumber: testBlockNumber })
   })
 
   it(`should execute an action, even if OperationStorage lock() was called by another address`, async () => {
     const openVaultAction = createAction(
-      await registry.getEntryHash(CONTRACT_NAMES.maker.OPEN_VAULT),
+      OpenVaultActionHash,
       [calldataTypes.maker.Open, calldataTypes.paramsMap],
       [
         {
@@ -50,15 +66,15 @@ describe.skip(`Reentrancy guard test | Unit`, async () => {
     )
 
     // LOCK OperationStorage before operation execution
-    await system.common.operationStorage.lock()
+    await system.OperationStorage.contract.lock()
 
     const [success] = await executeThroughProxy(
-      system.common.userProxyAddress,
+      userProxy.address,
       {
-        address: system.common.operationExecutor.address,
-        calldata: system.common.operationExecutor.interface.encodeFunctionData('executeOp', [
+        address: system.OperationExecutor.contract.address,
+        calldata: system.OperationExecutor.contract.interface.encodeFunctionData('executeOp', [
           [openVaultAction],
-          OPERATION_NAMES.common.CUSTOM_OPERATION, //just to skip operation's actions verification
+          OPERATION_NAME,
         ]),
       },
       signer,
@@ -66,7 +82,7 @@ describe.skip(`Reentrancy guard test | Unit`, async () => {
 
     expect(success).to.be.eq(true)
 
-    const vault = await getLastVault(provider, signer, system.common.userProxyAddress)
+    const vault = await getLastVault(provider, signer, userProxy.address)
 
     const cdpManagerContract = new ethers.Contract(
       ADDRESSES[Network.MAINNET].maker.common.CdpManager,
@@ -74,6 +90,6 @@ describe.skip(`Reentrancy guard test | Unit`, async () => {
       provider,
     ).connect(signer)
     const vaultOwner = await cdpManagerContract.owns(vault.id)
-    expect.toBeEqual(vaultOwner, system.common.userProxyAddress)
+    expect.toBeEqual(vaultOwner, userProxy.address)
   })
 })
