+++ packages/dma-contracts/utils/recursive-partial.ts
@@ -1,0 +1,8 @@
+// Recursive partial type for partials of nested objects
+export type RecursivePartial<T> = {
+  [P in keyof T]?: T[P] extends (infer U)[]
+    ? RecursivePartial<U>[]
+    : T[P] extends object
+    ? RecursivePartial<T[P]>
+    : T[P]
+}
